Add stable_diffusion train.sh

+++ models/map_model.py
@@ -35,7 +35,7 @@ def map_model(model_type):
         repo = "Bingsu/my-korean-stable-diffusion-v1-5"
         euler_ancestral_scheduler = EulerAncestralDiscreteScheduler.from_config(repo, subfolder="scheduler")
         pipe = StableDiffusionPipeline.from_pretrained(
-            repo, scheduler=euler_ancestral_scheduler, torch_dtype=torch.float16,
+            repo, scheduler=euler_ancestral_scheduler, torch_dtype=torch.float32,
         )
         pipe.to("cuda" if torch.cuda.is_available() else "cpu")
